mcp-claude-runner/requirements.txt
mcp[cli]
fastapi
uvicorn[standard]
python-jose[cryptography]
docker
